refactor(session): move display contracts out of services
Related to #434
