+++ apps/desktop/src/contracts/index.ts
@@ -2,5 +2,8 @@
 
 export { AiError, AiErrorCode } from './agentErrors';
 export type * from './attachments';
+export type * from './builtInToolPresentation';
 export type * from './json';
+export type * from './sessionStatus';
 export type * from './tooling';
+export type * from './widgets';
