+++ apps/desktop/src/contracts/sessionStatus.ts
@@ -1,0 +1,3 @@
+// Copyright (c) 2026. Qian Cheng. Licensed under GPL v3
+
+export type SessionStatusReminderKind = 'completed' | 'failed' | 'waiting_approval';
