+++ apps/desktop/src/types/session.ts
@@ -3,12 +3,12 @@
 import type { JSONContent } from '@tiptap/core';
 
 import type { AttachmentIndex as Index } from '@/contracts/attachments';
-import type { ToolExecutionSource as AiToolExecutionSource } from '@/contracts/tooling';
 import type {
     BuiltInToolConversationPresentation,
     BuiltInToolConversationSemantic,
-} from '@/services/BuiltInToolService';
-import type { ShowWidgetPayload } from '@/services/BuiltInToolService/tools/widgetTool';
+} from '@/contracts/builtInToolPresentation';
+import type { ToolExecutionSource as AiToolExecutionSource } from '@/contracts/tooling';
+import type { ShowWidgetPayload } from '@/contracts/widgets';
 
 /**
  * SearchView 与 agent 运行时共享的会话展示模型。
